Fix level dependency of "hzappp" feature extraction
The main reason for this update to the code is an update to the feature
extraction. The originally published implementation divided the log amplitude
values of the log Mel-spectrogram by the (frequency-dependet) values of the
level uncertainty before adding random values drawn from a standard normal
distribution. Although this seemed like an elegant way of removing the
information in frequency regions with high level uncertainty, it resulted
in an undesired level dependency of the effect of the level uncertainty.
This effect was observed in an attempt to simulate the level dependency
of the SRT (Plomp curves).

This release comes with an alternative implementation, where the values
of the log Mel-spectrogram are no longer divided by the values of the
level uncertainty, but the values drawn from the standard normal distribution
are multiplied by the values of the level uncertainty, which ocurrs due to
subsequent across-frequency integration by the Gabor filter banks. This change
avoids the undesired level dependency of the effect of the level uncertainty,
with only very little change to the simulation results of this study.

The main change is (now using singleton dimension expansion):
- log_melspec = bsxfun(@times,log_melspec,1./ul_mel) + randn(size(log_melspec));
+ log_melspec = log_melspec + ul_mel.*randn(size(log_melspec));

The simulation scripts also have been updated to run with Ubuntu 20.10 and
FADE 2.4.0. All simulations were re-run, the simulation results are also
part of this commit (ul2tintable.txt and matrix_simulated_data.txt).
The differences to the published results are minor.

Further changes:
- Use ramdisk (/dev/shm/) to run simulations (needs a size of at least ~24Gb)
- Updated evaluation scripts to run with Ubuntu 20.10